requirements.txt
python-gitlab==1.3.0
requests==2.18.4
six==1.11.0
iso8601
matplotlib
